Move domain redirect to controllers
