+++ controllers/domain-redirect.js
